Building Proper File Structure for DB
